+++ cmd/clean.go
@@ -4,10 +4,9 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
-	"tfgen/config"
+	"tfgen/tfgen"
 
 	"github.com/rs/zerolog/log"
-
 	"github.com/spf13/cobra"
 )
 
@@ -36,7 +35,7 @@ func clean(targetDir string) error {
 	if err != nil {
 		log.Fatal().Err(err).Msg("failed to get absolute path")
 	}
-	configHandler := config.NewConfigHandler(absTargetDir)
+	configHandler := tfgen.NewConfigHandler(absTargetDir)
 	if err := configHandler.ParseConfigFiles(); err != nil {
 		return err
 	}
