+++ tfgen/finder.go
@@ -1,4 +1,4 @@
-package config
+package tfgen
 
 import (
 	"fmt"
